requirements.txt
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
streamlit==1.32.0
plotly==5.19.0
python-dotenv==1.0.1
